+++ processor/src/test/fixtures/input/com/example/component/ProtectedCtorModel.java
@@ -1,0 +1,12 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+
+@ArezComponent( allowEmpty = true )
+abstract class ProtectedCtorModel
+{
+  @SuppressWarnings( "WeakerAccess" )
+  protected ProtectedCtorModel()
+  {
+  }
+}