+++ processor/src/test/fixtures/input/com/example/component/Suppressed1ProtectedCtorModel.java
@@ -1,0 +1,12 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+
+@ArezComponent( allowEmpty = true )
+abstract class Suppressed1ProtectedCtorModel
+{
+  @SuppressWarnings( "Arez:ProtectedConstructor" )
+  protected Suppressed1ProtectedCtorModel()
+  {
+  }
+}