+++ processor/src/test/fixtures/input/com/example/component/Suppressed2ProtectedCtorModel.java
@@ -1,0 +1,13 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.SuppressArezWarnings;
+
+@ArezComponent( allowEmpty = true )
+abstract class Suppressed2ProtectedCtorModel
+{
+  @SuppressArezWarnings( "Arez:ProtectedConstructor" )
+  protected Suppressed2ProtectedCtorModel()
+  {
+  }
+}