+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Add two new hook methods that are invoked after an `@Inverse` reference is added to a component or before an `@Inverse` reference is removed from a component. These hook methods are designated using the new annotations `@PreInverseRemove` and `@PostInverseAdd`. See the javadocs for full details on how to use the annotations.
+* Generate a suppressible warning if a protected constructor appears on an Arez component. The constructor should be either package access or public.
 
 ### [v0.159](https://github.com/arez/arez/tree/v0.159) (2019-12-17) · [Full Changelog](https://github.com/arez/arez/compare/v0.158...v0.159) · [API Differences](https://arez.github.io/api-diff?key=arez&old=0.158&new=0.159)